qmp: Add an implementation wrapper for qmp_drive_backup
We'd like to be able to specify the callback given to backup_start
manually in the case of transactions, so split apart qmp_drive_backup
into an implementation and a wrapper.

Switch drive_backup_prepare to use the new wrapper, but don't overload
the callback and closure yet.

This is kind of gross, but we need to forward-declare the wrapper for
the drive_backup transaction to be able to use. I could try moving
the transaction blocks lower instead, but I wanted (for v1, at least)
to minimize code movement so that the core changes were easiest to see.

If anyone has suggestions on organization, please suggest them.