ahci: correct PIO FIS semantics
This one is a little wonky.

Currently, we post our "PIO Setup FIS" before the actual transaction.
Odd.

Apparently what is supposed to happen, according to SATA 3.2 section
10.5.11 ("PIO Setup - Device to Host FIS") is:

- drive receives a command to read or write PIO data
- drive /prepares/ a PIO Setup FIS
- drive /prepares/ a data FIS
- drive /updates/ the Setup FIS with the register values after preparing
  the data FIS
- drive sends Setup FIS
- drive sends Data FIS

At least, that's my interpretation of what would need to happen in order
to have both "pre" and "post" register status values.

So what this patch does is:
- Generate most of the PIO Setup FIS prior to the data transfer
- Generate the rest upon completion

As a consequence, because we are mimicking metadata that accompanies each
data FIS for a PIO transfer, the number of interrupts will increase from
one per command to one per /transfer/, but this will not impact guests that
do not select to receive these spurious interrupts by leaving the PSE bit
unset.