ide: unify io_buffer_offset increments
This variable was originally added to keep track of offsets in AHCI
rather exclusively, but it was added to IDEState. AHCI fakes all PIO
transfers using DMA and a scatter-gather list. When the core or atapi
layers invoke HBA-specific mechanisms for transfers, they do not always
know that it is being backed by DMA or a sglist, so this offset is not
always updated by all the code everywhere.

If we modify it in dma_buf_commit, however, any HBA that needs to use
this offset to manage operating on only part of a sglist will have
access to it.

As a result, this will fix ATAPI PIO transfers performed through the AHCI
HBA, which were previously not modifying this value appropriately. This
should fix AHCI/ATAPI with OVMF, which tries to use these PIO transfers.

Reported-by: Hannes Reinecke <hare@suse.de>