configure: workaround for Clang 3.5.0
Clang++ 3.5 on Fedora 22 appears to have difficulty tolerating
D_FORTIFY_SOURCE for certain glibc headers, such as stdio.

This interferes, currently, with any arm target build.

Work around this by disabling FORTIFY_SOURCE for clang builds
if a problem is observed.

Newer versions of clang such as 3.5.2 (As seen in debian-testing)
or 3.7.0 (As seen in Fedora 23 Beta) are unaffected and will not
trigger this workaround.