iotests: 013: fix divide-by-zero to silence clang
iotest 013 will fail under Clang as of late because some of the
operations complete so quickly that t2=t2-t1 will actually be 0. When we
try to calculate ops/sec, we hit a divide by zero which clang is not too
happy about under -fsanitize=undefined.

Return "inf" in these cases when we lack the necessary precision to give
a real answer.