qcow2-bitmap: fail when removing bitmaps if nb_bitmaps is zero
Originally, we permitted qcow2_remove_persistent_dirty_bitmap to
silently fail if we tried to remove a bitmap from a qcow2 file that
was thought to have no bitmaps.

Now that we are caching the bitmap list, it ought not matter if we
have flushed to disk yet or not, so this case SHOULD report an error.