tests/acceptance: Don't wait for cubieboard
cubieboard does not have a functioning reboot, it halts and QEMU does
not exit. We can't wait for the guest in this case, so allow the test to
call the implicit vm.shutdown() instead which does not wait for the
guest's cooperation.

Attempting the reboot and omittimg the vm.wait() may introduce race
conditions if and when the reboot is ever improved, so removing the
reboot test for cubieboard helps avoid a possibly difficult debugging
session some time in the future.